# starter/gin/docs/swagger.yaml
basePath: /api
definitions:
...
  org.CreateOrgDTO:
    properties:
      avatar:
        maxLength: 255
        type: string
      description:
        maxLength: 500
        type: string
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
      - display_name
      - name
    type: object
...
  org.OrgDTO:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  org.UpdateOrgDTO:
    properties:
      avatar:
        maxLength: 255
        type: string
      description:
        maxLength: 500
        type: string
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      status:
        enum:
          - active
          - suspended
        type: string
    type: object
...
  response.DataResponse-org_OrgDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
          - $ref: "#/definitions/org.OrgDTO"
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: "#/definitions/response.ErrorDetail"
    type: object
  response.MessageResponse:
    description: 纯消息响应格式
    properties:
      code:
        description: HTTP 状态码
        type: integer
      message:
        description: 消息描述
        type: string
    type: object
...
  response.PagedResponse-org_OrgDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 数据列表
        items:
          $ref: "#/definitions/org.OrgDTO"
        type: array
      message:
        description: 消息描述
        type: string
      meta:
        allOf:
          - $ref: "#/definitions/response.PaginationMeta"
        description: 分页信息
    type: object
...
  response.PaginationMeta:
    properties:
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      per_page:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
      warning:
        description: 页码越界警告
        type: string
    type: object
...
host: localhost:40012
info:
  contact:
    email: lwmacct@icloud.com
    name: API Support
    url: https://github.com/lwmacct/260101-go-pkg-ddd/issues
  description: 基于 DDD+CQRS 架构的 Go Web 应用 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lwmacct/260101-go-pkg-ddd
  title: Go DDD Template API
  version: 1.0.3
paths:
...
  /api/admin/orgs:
    get:
      consumes:
        - application/json
      description: 分页获取所有组织
      parameters:
        - default: 20
          description: Limit 每页数量，默认 20，最大 1000
          in: query
          maximum: 1000
          minimum: 1
          name: limit
          type: integer
        - default: 1
          description: Page 页码，从 1 开始
          in: query
          minimum: 1
          name: page
          type: integer
        - enum:
            - active
            - inactive
          in: query
          name: status
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: 组织列表
          schema:
            $ref: "#/definitions/response.PagedResponse-org_OrgDTO"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 组织列表
      tags:
        - Admin - Organization Management
    post:
      consumes:
        - application/json
      description: 系统管理员创建新组织
      parameters:
        - description: 组织信息
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/org.CreateOrgDTO"
      produces:
        - application/json
      responses:
        "201":
          description: 组织创建成功
          schema:
            $ref: "#/definitions/response.DataResponse-org_OrgDTO"
        "400":
          description: 参数错误或组织名已存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 创建组织
      tags:
        - Admin - Organization Management
  /api/admin/orgs/{id}:
    delete:
      consumes:
        - application/json
      description: 软删除组织
      parameters:
        - description: 组织ID
          in: path
          minimum: 1
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 组织删除成功
          schema:
            $ref: "#/definitions/response.MessageResponse"
        "400":
          description: 无效的组织ID
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "404":
          description: 组织不存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 删除组织
      tags:
        - Admin - Organization Management
    get:
      consumes:
        - application/json
      description: 根据 ID 获取组织详情
      parameters:
        - description: 组织ID
          in: path
          minimum: 1
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 组织详情
          schema:
            $ref: "#/definitions/response.DataResponse-org_OrgDTO"
        "400":
          description: 无效的组织ID
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "404":
          description: 组织不存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 组织详情
      tags:
        - Admin - Organization Management
    put:
      consumes:
        - application/json
      description: 更新组织信息
      parameters:
        - description: 组织ID
          in: path
          minimum: 1
          name: id
          required: true
          type: integer
        - description: 更新信息
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/org.UpdateOrgDTO"
      produces:
        - application/json
      responses:
        "200":
          description: 组织更新成功
          schema:
            $ref: "#/definitions/response.DataResponse-org_OrgDTO"
        "400":
          description: 参数错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "404":
          description: 组织不存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 更新组织
      tags:
        - Admin - Organization Management
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"